SprintTwo/requirements.txt
-e .
aws-cdk.core==1.122.0
aws-cdk.aws_lambda==1.122.0
aws-cdk.aws_events==1.122.0
aws-cdk.aws_iam==1.122.0
aws-cdk.aws_events_targets==1.122.0
boto3